+++ src/main/java/org/hydev/mcpm/client/uninstall/UninstallInput.java
@@ -1,4 +1,4 @@
-package org.hydev.mcpm.client.uninstallUseCase;
+package org.hydev.mcpm.client.uninstall;
 
 import org.hydev.mcpm.client.database.inputs.SearchPackagesType;
 import org.hydev.mcpm.client.injector.PluginLoader;
@@ -9,13 +9,13 @@ import org.hydev.mcpm.client.injector.PluginLoader;
  * @param name Plugin name
  * @param ulb PluginLoader object
  * @param spType SearchPackagesType
- * @param remove_depen remove dependencies or not
+ * @param recursive remove dependencies or not
  */
 public  record UninstallInput(
     String name,
     PluginLoader ulb,
     SearchPackagesType spType,
-    boolean remove_depen
+    boolean recursive
 ) { }
